ibkr_bots/bots/bot_B_buywrite.py
#!/usr/bin/env python3
"""
Bot B: Micro Buy-Write Strategy (SKELETON)

TODO: Implement covered call / buy-write strategy for day trading.
"""

import logging
from typing import Dict, Any

logger = logging.getLogger('BotB')


class BuyWriteBot:
    """
    TODO: Buy-Write (Covered Call) trading bot
    
    Strategy Overview:
    1. Buy small equity positions during favorable conditions
    2. Sell calls against 25-50% of position
    3. Target 40-60% premium capture
    4. Close before market close
    
    Risk Management:
    - Position sizing based on volatility
    - Delta hedging as needed
    - Time-based exits
    """
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        logger.warning("Bot B (Buy-Write) is a SKELETON - not implemented")
        
    def evaluate_entry_conditions(self, symbol: str) -> Dict[str, Any]:
        """
        TODO: Evaluate buy-write entry conditions
        
        Should consider:
        - Underlying momentum/direction
        - IV levels for call premium
        - Time to expiration
        - Position size limits
        """
        logger.info("TODO: Implement buy-write entry evaluation")
        return {
            'symbol': symbol,
            'entry_allowed': False,
            'reason': 'Bot B not yet implemented'
        }
    
    def enter_position(self, symbol: str, evaluation: Dict[str, Any]) -> bool:
        """
        TODO: Enter buy-write position
        
        Steps:
        1. Buy underlying shares
        2. Sell call options against position
        3. Set up monitoring and exit rules
        """
        logger.info("TODO: Implement buy-write position entry")
        return False
    
    def manage_positions(self):
        """
        TODO: Manage existing buy-write positions
        
        Monitor for:
        - Time decay benefits
        - Delta hedging needs
        - Early assignment risk
        - Profit target achievement
        """
        logger.debug("TODO: Implement buy-write position management")
    
    def run(self):
        """
        TODO: Main execution loop for buy-write strategy
        """
        logger.info("Bot B (Buy-Write) started in skeleton mode")
        logger.warning("⚠️  This bot is not yet implemented - no trades will be executed")
        
        # Skeleton main loop
        while True:
            try:
                # TODO: Add actual trading logic here
                logger.debug("Bot B skeleton - no operations")
                break  # Exit for now
                
            except KeyboardInterrupt:
                logger.info("Bot B shutdown requested")
                break
            except Exception as e:
                logger.error(f"Bot B error: {e}")
                break


def main():
    """Entry point for Bot B"""
    import argparse
    import yaml
    
    parser = argparse.ArgumentParser(description='Buy-Write Options Trading Bot')
    parser.add_argument('--config', default='config/strategy.yaml', help='Config file path')
    
    args = parser.parse_args()
    
    try:
        with open(args.config, 'r') as f:
            config = yaml.safe_load(f)
    except Exception as e:
        logger.error(f"Failed to load config: {e}")
        return
    
    bot = BuyWriteBot(config)
    bot.run()


if __name__ == '__main__':
    main()

