ibkr_bots/requirements.txt
ib_insync>=0.9.86
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
streamlit>=1.28.0